+++ packages/bluesky_text/lib/src/extractor/markdown_extractor.dart
@@ -3,16 +3,21 @@
 // modification, are permitted provided the conditions.
 
 // 🌎 Project imports:
-import 'bluesky_text.dart';
-import 'entities/byte_indices.dart';
-import 'entities/markdown/markdown_link_entity.dart';
-import 'regex/markdown_link.dart';
-import 'regex/valid_mention.dart';
-import 'regex/valid_url.dart';
-import 'unicode_string.dart';
-import 'utils.dart';
+import '../bluesky_text.dart';
+import '../entities/byte_indices.dart';
+import '../entities/markdown/markdown_link_entity.dart';
+import '../regex/markdown_link.dart';
+import '../regex/valid_ascii_domain.dart';
+import '../regex/valid_hashtag.dart';
+import '../regex/valid_mention.dart';
+import '../regex/valid_url.dart';
+import '../unicode_string.dart';
+import '../utils.dart';
 
 const markdownLinksExtractor = MarkdownLinksExtractor();
+
+/// `[`, `]`, `(`, `)`
+const _parensCount = 4;
 
 final class MarkdownLinksExtractor {
   const MarkdownLinksExtractor();
@@ -24,7 +29,20 @@ final class MarkdownLinksExtractor {
 
     for (final match in markdownLinkRegex.allMatches(text.value)) {
       final linkText = match.markdownLinkText;
-      final linkUrl = match.markdownLinkUrl;
+      final urlMatch = validUrlRegex.firstMatch(
+        text.value.substring(match.start),
+      );
+
+      if (urlMatch == null) continue;
+
+      final protocol = urlMatch.protocol;
+      final domain = urlMatch.domain;
+      final portNumber = urlMatch.portNumber;
+      final urlPath = urlMatch.path;
+      final urlQuery = urlMatch.query;
+
+      final linkUrl = '$protocol${getFirstValidDomain(domain)}'
+          '$portNumber$urlPath$urlQuery';
 
       if (!_isValidMarkdownLink(linkText, linkUrl)) continue;
 
@@ -35,7 +53,7 @@ final class MarkdownLinksExtractor {
           indices: ByteIndices(
             start: text.value.toUtf8Index(match.start),
             end: text.value.toUtf8Index(
-              match.start + match.markdownLink.length,
+              match.start + linkText.length + linkUrl.length + _parensCount,
             ),
           ),
         ),
@@ -51,6 +69,8 @@ final class MarkdownLinksExtractor {
     if (Uri.tryParse(url) == null) return false;
 
     //* Prevent users from linking to specific mentions text for their safety.
-    return !validMentionRegex.hasMatch(text) && validUrlRegex.hasMatch(url);
+    return !validMentionRegex.hasMatch(text) &&
+        !validHashtagRegex.hasMatch(text) &&
+        !validUrlRegex.hasMatch(text);
   }
 }